fix: a lot of problem

+++ backend/src/main/java/com/example/demo/service/TagService.java
@@ -3,9 +3,17 @@ package com.example.demo.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.example.demo.model.Tag;
-import com.example.demo.repository.TagRepository;
+import javax.servlet.http.HttpServletRequest;
+
 import com.example.demo.model.ResourceNotFoundException;
+import com.example.demo.model.Tag;
+
+import com.example.demo.service.JwtService;
+import com.example.demo.service.UserService;
+
+import com.example.demo.dto.TagResult;
+
+import com.example.demo.repository.TagRepository;
 
 import java.util.List;
 
@@ -13,27 +21,70 @@ import java.util.List;
 public class TagService {
     @Autowired
     private TagRepository tagRepository;
+    
+    @Autowired
+    private JwtService jwtService;
+
+    @Autowired
+    private UserService userService;
 
     public List<Tag> getAllTags() {
         return tagRepository.findAll();
     }
 
-    public Tag addTag(Tag tag) {
-        return tagRepository.save(tag);
+    public TagResult createTag(Tag tag, HttpServletRequest request) {
+        String token = request.getHeader("Authorization").substring(7);
+        String userId = jwtService.extractUserId(token);
+        User user = userService.findById(userId);
+        TagResult result = new TagResult();
+
+        if (user.getPermission() == 0) {
+            result.setResultCode(1);
+        } else {
+            result.setTag(tagRepository.save(tag));
+            result.setResultCode(0);
+        }
+
+        return result;
     }
 
-    public void deleteTag(Integer id) {
+    public int deleteTag(Integer id, HttpServletRequest request) {
+        String token = request.getHeader("Authorization").substring(7);
+        String userId = jwtService.extractUserId(token);
+        User user = userService.findById(userId);
+
         Tag tag = getTag(id);
-        tagRepository.delete(tag);
+
+        if (user.getPermission() == 0) {
+            return 1;
+        } else if (tag == null) {
+            return 2;
+        } else {
+            tagRepository.delete(tag);
+            return 0
+        }
     }
 
-    public Tag updateTag(Integer id, Tag tagDetails) {
+    public TagResult updateTag(Integer id, Tag tagDetails, HttpServletRequest request) {
+        String token = request.getHeader("Authorization").substring(7);
+        String userId = jwtService.extractUserId(token);
+        User user = userService.findById(userId);
+
         Tag tag = getTag(id);
-        tag.setTagName(tagDetails.getTagName());
-        return tagRepository.save(tag);
-    }
+        TagResult result = new TagResult();
 
-    private Tag getTag(Integer id) {
-        return tagRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Tag", "id", id));
+        if (user.getPermission() == 0) {
+            result.setResultCode(1);
+        } else if (tag == null) {
+            result.setResultCode(2);
+        } else{
+            tag.setTagName(tagDetails.getTagName());
+            tag.setTagAbbr(tagDetails.getTagAbbr());
+
+            result.setTag(tagRepository.save(tag));
+            result.setResultCode(0);
+        }
+
+        return result;
     }
 }
